Create a new demo project: demo-jquery-table. (#212)
In this demo, Hummingbird is used to create a server-side wrapper over a
table rendered with jQuery. The server creates the table and invokes the
initialization methods needed for the jQuery plugin, and the table
plugin does its job, providing sorting, filtering and grouping features. The selection of table rows is handled by the server logic.

This is not a full featured demo - it only show how can an arbitrary
component can be handled from the server side by using Hummingbird.

Fixes #212.